+++ src/Entities/Adjustment.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 namespace Paddle\SDK\Entities;
 
 use Paddle\SDK\Entities\Adjustment\AdjustmentItem;
+use Paddle\SDK\Entities\Adjustment\AdjustmentTaxRatesUsed;
 use Paddle\SDK\Entities\Shared\Action;
 use Paddle\SDK\Entities\Shared\AdjustmentStatus;
 use Paddle\SDK\Entities\Shared\AdjustmentTotals;
@@ -21,7 +22,8 @@ use Paddle\SDK\Entities\Shared\PayoutTotalsAdjustment;
 class Adjustment implements Entity
 {
     /**
-     * @param array<AdjustmentItem> $items
+     * @param array<AdjustmentItem>         $items
+     * @param array<AdjustmentTaxRatesUsed> $taxRatesUsed
      */
     private function __construct(
         public string $id,
@@ -36,6 +38,7 @@ class Adjustment implements Entity
         public array $items,
         public AdjustmentTotals $totals,
         public PayoutTotalsAdjustment|null $payoutTotals,
+        public array $taxRatesUsed,
         public \DateTimeInterface $createdAt,
         public \DateTimeInterface|null $updatedAt,
     ) {
@@ -56,6 +59,7 @@ class Adjustment implements Entity
             items: array_map(fn (array $item): AdjustmentItem => AdjustmentItem::from($item), $data['items']),
             totals: AdjustmentTotals::from($data['totals']),
             payoutTotals: isset($data['payout_totals']) ? PayoutTotalsAdjustment::from($data['payout_totals']) : null,
+            taxRatesUsed: array_map(fn (array $taxRateUsed): AdjustmentTaxRatesUsed => AdjustmentTaxRatesUsed::from($taxRateUsed), $data['tax_rates_used'] ?? []),
             createdAt: DateTime::from($data['created_at']),
             updatedAt: DateTime::from($data['updated_at']),
         );