+++ src/Entities/Shared/AvailablePaymentMethods.php
@@ -20,7 +20,10 @@ use Paddle\SDK\PaddleEnum;
  * @method static AvailablePaymentMethods Card()
  * @method static AvailablePaymentMethods GooglePay()
  * @method static AvailablePaymentMethods Ideal()
+ * @method static AvailablePaymentMethods Offline()
  * @method static AvailablePaymentMethods Paypal()
+ * @method static AvailablePaymentMethods Unknown()
+ * @method static AvailablePaymentMethods WireTransfer()
  */
 final class AvailablePaymentMethods extends PaddleEnum
 {
@@ -30,5 +33,8 @@ final class AvailablePaymentMethods extends PaddleEnum
     private const Card = 'card';
     private const GooglePay = 'google_pay';
     private const Ideal = 'ideal';
+    private const Offline = 'offline';
     private const Paypal = 'paypal';
+    private const Unknown = 'unknown';
+    private const WireTransfer = 'wire_transfer';
 }