+++ src/Notifications/Entities/Transaction.php
@@ -52,7 +52,6 @@ class Transaction implements Entity
         public \DateTimeInterface $createdAt,
         public \DateTimeInterface $updatedAt,
         public \DateTimeInterface|null $billedAt,
-        public string|null $receiptData,
     ) {
     }
 
@@ -81,7 +80,6 @@ class Transaction implements Entity
             createdAt: DateTime::from($data['created_at']),
             updatedAt: DateTime::from($data['updated_at']),
             billedAt: isset($data['billed_at']) ? DateTime::from($data['billed_at']) : null,
-            receiptData: $data['receipt_data'] ?? null,
         );
     }
 }